# docs/swagger.yaml
definitions:
...
  entity.SysApi:
    properties:
      apiGroup:
        description: api组
        type: string
      create_time:
        type: string
      description:
        description: api中文描述
        type: string
      id:
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      update_time:
        type: string
    required:
    - path
    type: object
...
  entity.SysUser:
    properties:
      avatar:
        description: 头像
        type: string
      create_time:
        type: string
      createBy:
        type: string
      deptId:
        description: 部门编码
        type: integer
      email:
        description: 邮箱
        type: string
      nickName:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      postId:
        description: 职位编码
        type: integer
      postIds:
        description: 多岗位
        type: string
      remark:
        description: 备注
        type: string
      roleId:
        description: 角色编码
        type: integer
      roleIds:
        description: 多角色
        type: string
      salt:
        description: 盐
        type: string
      sex:
        description: 性别
        type: string
      status:
        type: string
      update_time:
        type: string
      updateBy:
        type: string
      userId:
        description: 编码
        type: integer
      username:
        type: string
    type: object
  entity.SysUserPwd:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
  form.Login:
    properties:
      captcha:
        description: 验证码
        type: string
      captchaId:
        description: 验证码ID
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /system/api:
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.SysApi'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - X-TOKEN: []
      summary: 创建基础api
      tags:
      - SysApi
  /system/api/{id}:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - X-TOKEN: []
      summary: 删除api
      tags:
      - SysApi
    get:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取api
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取api
      tags:
      - SysApi
...
  /system/dict/type/export:
    get:
      description: 导出数据
      parameters:
      - description: userName
        in: query
        name: userName
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      - description: status
        in: query
        name: status
        type: string
      responses:
        "200":
          description: '{"code": 400, "message": "删除失败"}'
          schema:
            type: string
      summary: 导出用户
      tags:
      - 用户
...
  /system/user/getInit:
    get:
      description: 获取JSON
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            type: string
      security:
      - "": []
      summary: 获取添加用户角色和职位
      tags:
      - 用户
  /system/user/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/form.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"登陆成功"}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - Base
...
  /system/user/profile:
    get:
      description: 获取JSON
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            type: string
      security:
      - "": []
      summary: 获取当前登录用户
      tags:
      - 个人中心
  /system/user/profileAvatar:
    post:
      description: 修改头像
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      summary: 修改头像
      tags:
      - 用户
...
  /system/user/sysUser:
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: 用户数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.SysUser'
      responses:
        "200":
          description: '{"code": 400, "message": "添加失败"}'
          schema:
            type: string
      summary: 创建用户
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: 用户数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.SysUser'
      responses:
        "200":
          description: '{"code": 400, "message": "添加失败"}'
          schema:
            type: string
      summary: 修改用户状态
      tags:
      - 用户
  /system/user/sysUser/{userId}:
    get:
      description: 获取JSON
      parameters:
      - description: 用户编码
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            type: string
      security:
      - "": []
      summary: 获取用户
      tags:
      - 用户
  /system/user/sysUserList:
    get:
      description: 获取JSON
      parameters:
      - description: userName
        in: query
        name: userName
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: 页条数
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageNum
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      security:
      - X-TOKEN: []
      summary: 列表数据
      tags:
      - 用户
  /system/user/sysuser/{userId}:
    delete:
      description: 删除数据
      parameters:
      - description: 多个id 使用逗号隔开
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 400, "message": "删除失败"}'
          schema:
            type: string
      summary: 删除用户数据
      tags:
      - 用户
  /system/user/updatePwd:
    post:
      description: 修改密码
      parameters:
      - description: pwd
        in: body
        name: pwd
        required: true
        schema:
          $ref: '#/definitions/entity.SysUserPwd'
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      summary: 修改密码
      tags:
      - 用户
swagger: "2.0"